% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poppy_api_key.R
\name{get_poppy_api_key}
\alias{get_poppy_api_key}
\title{Retrieve the Poppy API key}
\usage{
get_poppy_api_key()
}
\value{
The value of \code{POPPY_API_KEY} or \code{NA_character_} when the variable is
not set.
}
\description{
Looks up \code{POPPY_API_KEY} in the environment for use with \code{poppy_chat()}.
}
\details{
Intended to be passed as the \code{credentials} argument to
\code{poppy_chat()}, following the \code{ellmer::chat_openai_compatible()}
convention of accepting a function that returns a key.
}
\examples{
Sys.setenv(POPPY_API_KEY = "abc123")
get_poppy_api_key()

}
